Some Refactoring And Commenting Done
